Handle group when blocking

+++ app/src/main/java/com/toshi/manager/SofaMessageManager.java
@@ -236,11 +236,13 @@ public final class SofaMessageManager {
                 .subscribeOn(Schedulers.io());
     }
 
+    //TODO: Support groups
     public Single<Conversation> rejectConversation(final Conversation conversation) {
+        if (conversation.isGroup()) return Single.error(new Throwable("Can't block groups"));
         return BaseApplication
                 .get()
                 .getRecipientManager()
-                .blockUser(conversation.getRecipient().getUser().getToshiId())
+                .blockUser(conversation.getThreadId())
                 .andThen(deleteConversation(conversation))
                 .toSingle(() -> conversation);
     }
